+++ lib/browser/browser.rb
@@ -28,6 +28,7 @@ require "browser/facebook"
 require "browser/otter"
 require "browser/instagram"
 require "browser/yandex_browser"
+require "browser/snapchat"
 
 require "browser/bot"
 require "browser/middleware"
@@ -58,6 +59,7 @@ module Browser
       Otter,
       Facebook,             # must be placed before Chrome and Safari
       Instagram,            # must be placed before Chrome and Safari
+      Snapchat,             # must be placed before Chrome and Safari
       Weibo,                # must be placed before Chrome and Safari
       QQ,                   # must be placed before Chrome and Safari
       Alipay,               # must be placed before Chrome and Safari